Add temp file

+++ weather.sh
@@ -2,13 +2,14 @@
 
 echo "Downloading weather data"
 mkdir -p ../data/weather
-curl -o data/weather/`date +"%Y%m%d_%H%M%S_athenry.json"` https://prodapi.metweb.ie/observations/athenry/today
-
+curl --ssl-no-revoke -o data/weather/`date +"%Y%m%d_%H%M%S_athenry.json"` https://prodapi.metweb.ie/observations/athenry/today
 echo "Download complete"
 
-# Keep only the last 10 JSON files in the directory
+# Keep only the last 10 JSON files in the directory if there are 11 or more
 echo "Cleaning up old JSON files"
-ls -tp data/weather/*.json | grep -v '/$' | tail -n +11 | xargs -d '\n' rm -- 2>/dev/null
+if [ $(ls -1 data/weather/*.json 2>/dev/null | wc -l) -ge 11 ]; then
+    ls -tp data/weather/*.json | grep -v '/$' | tail -n +11 | xargs -d '\n' rm -- 2>/dev/null
+fi
 
 echo "Old JSON files removed"
